+++ cppcheck.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 macros=($(gcc -E -dM - < /dev/null | sed -E 's/^#define (\w+).*/-D\1/g'))
 
-cppcheck --enable=all --std=c11 --quiet "${macros[@]}"\
+cppcheck\
+	--enable=all --disable=style,unusedFunction\
+	--std=c11 --quiet -j 2\
+	"${macros[@]}"\
 	-I/usr/include/\
 	-I/usr/include/linux/\
 	-I/usr/include/x86_64-linux-gnu/\