+++ src/gml-glitch-logo/module.h
@@ -1,19 +1,25 @@
 /**
  * Заголовок модуля. Содержит объявления функций, которые реализует модуль.
  */
-#ifndef GML_PIXELS_MODULE_H
-#define GML_PIXELS_MODULE_H
+ 
+#ifndef GML_GLITCH_LOGO_MODULE_H
+#define GML_GLITCH_LOGO_MODULE_H
 
 #include "common.h"
-#include "util/bitset2d.h"
 #include <libpng/png.h>
+#include <freetype2/ft2build.h>
+#include FT_FREETYPE_H
 
-#define PIXEL_SIZE 4
+#define FPS 8
+#define GLYPH_HEIGHT 20
 
+extern color_t text_color;
+extern int RANDOM_CONSTANT;
+extern const char* system_name;
 extern png_structp png_ptr;
 extern png_infop info_ptr, end_info;
-
-extern bitset2d buffer;
+extern FT_Face face;
+extern color_t* bg_buffer; // Буфер размером height
 
 // Следующие функции вызываются в том порядке, в котором объявлены
 
@@ -22,11 +28,7 @@ void gml_read_config(config_t*);
 /** Загружает и инициализирует ресурсы модуля */
 void gml_setup(void);
 
-/**
- * Загружает и инициализирует ресурсы модуля после загрузки libdrm.
- * @param width - ширина экрана.
- * @param height - высота экрана.
- */
+/** Загружает и инициализирует ресурсы модуля после загрузки libdrm */
 void gml_setup_after_drm(uint16_t width, uint16_t height);
 
 
@@ -41,10 +43,10 @@ void gml_setup_after_drm(uint16_t width, uint16_t height);
 void gml_draw(int tick, uint16_t width, uint16_t height, color_t* frame, double supposed_time);
 
 
-/** Освобождает ресурсы модуля перед освобождением ресурсов libdrm. */
+/** Освобождает ресурсы модуля перед освобождением ресурсов libdrm */
 void gml_cleanup_before_drm(void);
 
-/** Освобождает ресурсы модуля. */
+/** Освобождает ресурсы модуля */
 void gml_cleanup(void);
 
 #endif